Rename MapErrorAsNext to OnErrorReturn
